refactor(backend): split upload session module by lifecycle phase
Split mediaAssets/uploadSessions/index.ts into shared, creation, reads, completion,
abort and writerAttempts modules, keeping index.ts as a barrel with an unchanged
exported surface so no caller outside the directory needs an edit.
